Updated functions/s2_scl_weights.js (added 1 line, deleted 1 line).

+++ functions/s2_scl_weights.js
@@ -25,7 +25,7 @@ exports.s2_scl_weights = function(img_coll, AOI, scale_to_use){
     .set({
       'system:time_start':time_start_value,
       'system:footprint':footprint
-    }).toInt();
+    });
     
     return mosaic;
   };
